refactor: 관리자 페이지 코드 리뷰 피드백 반영
- 신고 관련글 Link 제거 (targetId 필드 미제공으로 인한 잘못된 라우팅 방지)
- 신고 처리 모달 SUSPEND 옵션 제거, PRIVATE 라벨 명확화
- categoryId falsy 체크 버그 수정 (0 값 스킵 방지)
- formatDate 유틸 함수 utils.ts로 추출 (4곳 중복 제거)
- 공지사항 내용 placeholder 컬럼 제거
- 카테고리 DnD 미구현 cursor-grab → cursor-default
